Rename lisp_lang.ts to lisp.ts
Update import in spec/specs.ts, package.json scripts/main, and CLAUDE.md.
